+++ src/maps/OccupancyGridClient.js
@@ -28,7 +28,7 @@ ROS2D.OccupancyGridClient = function(options) {
   this.currentGrid = new createjs.Shape();
   this.rootObject.addChild(this.currentGrid);
   // work-around for a bug in easeljs -- needs a second object to render correctly
-  this.rootObject.addChild(new ROS2D.Grid({size:0}));
+  this.rootObject.addChild(new ROS2D.Grid({size:1}));
 
   // subscribe to the topic
   var rosTopic = new ROSLIB.Topic({